requirements.txt
flet
sqlalchemy
psycopg2-binary
pandas
openpyxl
python-dotenv
